Created wallet created event and moved some packages
