+++ src/main/kotlin/com/montebruni/holder/account/domain/events/data/UserEventData.kt
@@ -1,4 +1,4 @@
-package com.montebruni.holder.account.events.user
+package com.montebruni.holder.account.domain.events.data
 
 import com.montebruni.holder.account.domain.entity.Status
 import com.montebruni.holder.account.domain.valueobject.Password
