+++ src/main/kotlin/com/montebruni/holder/account/usecase/CreateCustomer.kt
@@ -1,10 +1,10 @@
 package com.montebruni.holder.account.usecase
 
 import com.montebruni.holder.account.domain.entity.Customer
+import com.montebruni.holder.account.domain.events.data.CustomerCreatedEvent
 import com.montebruni.holder.account.domain.exception.UserNotFoundException
 import com.montebruni.holder.account.domain.port.CustomerRepository
 import com.montebruni.holder.account.domain.port.UserRepository
-import com.montebruni.holder.account.events.customer.CustomerCreatedEvent
 import com.montebruni.holder.account.usecase.input.CreateCustomerInput
 import com.montebruni.holder.account.usecase.input.toCustomer
 import com.montebruni.holder.account.usecase.output.CreateCustomerOutput
