+++ src/main/kotlin/com/montebruni/holder/account/usecase/CreateUser.kt
@@ -1,9 +1,9 @@
 package com.montebruni.holder.account.usecase
 
 import com.montebruni.holder.account.domain.entity.User
+import com.montebruni.holder.account.domain.events.data.UserCreatedEvent
 import com.montebruni.holder.account.domain.exception.UserAlreadyExistsException
 import com.montebruni.holder.account.domain.port.UserRepository
-import com.montebruni.holder.account.events.user.UserCreatedEvent
 import com.montebruni.holder.account.usecase.input.CreateUserInput
 import com.montebruni.holder.account.usecase.input.toUser
 import com.montebruni.holder.account.usecase.output.CreateUserOutput
