+++ src/test/kotlin/com/montebruni/holder/account/domain/events/listener/UserMailListenerTest.kt
@@ -1,4 +1,4 @@
-package com.montebruni.holder.account.events.user
+package com.montebruni.holder.account.domain.events.listener
 
 import com.montebruni.holder.account.domain.mail.UserCreatedMailData
 import com.montebruni.holder.account.domain.mail.UserCreatedMailTemplate
