+++ src/test/kotlin/com/montebruni/holder/account/usecase/CreateUserTest.kt
@@ -1,9 +1,9 @@
 package com.montebruni.holder.account.usecase
 
 import com.montebruni.holder.account.domain.entity.User
+import com.montebruni.holder.account.domain.events.data.UserCreatedEvent
 import com.montebruni.holder.account.domain.exception.UserAlreadyExistsException
 import com.montebruni.holder.account.domain.port.UserRepository
-import com.montebruni.holder.account.events.user.UserCreatedEvent
 import com.montebruni.holder.common.event.EventPublisher
 import com.montebruni.holder.configuration.UnitTests
 import com.montebruni.holder.fixtures.createUser
