+++ src/test/kotlin/com/montebruni/holder/configuration/BaseRestIT.kt
@@ -1,6 +1,7 @@
 package com.montebruni.holder.configuration
 
 import com.fasterxml.jackson.databind.ObjectMapper
+import com.montebruni.holder.infrastructure.configuration.JacksonConfiguration
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Import
