+++ src/test/kotlin/com/montebruni/holder/fixtures/UserFixtures.kt
@@ -1,9 +1,9 @@
 package com.montebruni.holder.fixtures
 
 import com.montebruni.holder.account.domain.entity.User
+import com.montebruni.holder.account.domain.events.data.UserCreatedEvent
 import com.montebruni.holder.account.domain.valueobject.Password
 import com.montebruni.holder.account.domain.valueobject.Username
-import com.montebruni.holder.account.events.user.UserCreatedEvent
 import com.montebruni.holder.account.infrastructure.database.postgres.model.UserPostgresModel
 import com.montebruni.holder.account.usecase.input.CreateUserInput
 import java.util.UUID
