+++ src/test/kotlin/com/montebruni/holder/fixtures/WalletFixtures.kt
@@ -1,7 +1,7 @@
 package com.montebruni.holder.fixtures
 
-import com.montebruni.holder.common.valueobject.Amount
 import com.montebruni.holder.wallet.domain.entity.Wallet
+import com.montebruni.holder.wallet.domain.valueobject.Amount
 import com.montebruni.holder.wallet.infrastructure.database.postgres.model.WalletPostgresModel
 import java.util.UUID
