+++ src/test/kotlin/com/montebruni/holder/wallet/domain/valueobject/AmountTest.kt
@@ -1,4 +1,4 @@
-package com.montebruni.holder.common.valueobject
+package com.montebruni.holder.wallet.domain.valueobject
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
